# openapi.yaml
openapi: 3.1.0
info:
  title: Java Spring Learning API
  description: "Learning project for core Java concepts, Spring Boot, Swagger UI,\
    \ and OpenAPI YAML."
  contact:
    name: Learning Repo
    url: https://example.com
  license:
    name: For learning use
    url: https://opensource.org/licenses/MIT
  version: v1
servers:
- url: http://127.0.0.1:9091
  description: Generated server url
tags:
- name: Students
  description: "Student examples for learning Spring Boot controllers, services, and\
    \ dependency injection."
- name: Courses
  description: "Database-backed examples for validation, JPA, H2, and security."
- name: Learning Info
  description: Configuration-properties example exposed as a small public endpoint.
paths:
  /api/courses/{id}:
    get:
      tags:
      - Courses
      summary: Get a course by id
      description: Returns one persisted course.
      operationId: findById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Course not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Course found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CourseResponse"
    put:
      tags:
      - Courses
      summary: Update a course
      description: Replaces an existing course using the validated request body.
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Course not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Course updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CourseResponse"
    delete:
      tags:
      - Courses
      summary: Delete a course
      description: Deletes a course by id.
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Course not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "204":
          description: Course deleted
  /api/students:
    get:
      tags:
      - Students
      summary: List students
      description: Returns the in-memory student list used by this learning project.
      operationId: findAll
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Students loaded successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentResponse"
    post:
      tags:
      - Students
      summary: Create a student
      description: Creates a new student from a request model and returns the response
        model.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentRequest"
        required: true
      responses:
        "400":
          description: Invalid request data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "201":
          description: Student created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StudentResponse"
  /api/courses:
    get:
      tags:
      - Courses
      summary: List courses
      description: Returns all persisted courses. This endpoint requires basic authentication.
      operationId: findAll_1
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Courses loaded successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseResponse"
    post:
      tags:
      - Courses
      summary: Create a course
      description: Validates the request body and stores a course in H2.
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
        required: true
      responses:
        "400":
          description: Validation failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "201":
          description: Course created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CourseResponse"
  /api/students/{id}:
    get:
      tags:
      - Students
      summary: Get a student by id
      description: Returns one student or a 404 error if the student does not exist.
      operationId: findById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Student not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Student found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StudentResponse"
  /api/learning-info:
    get:
      tags:
      - Learning Info
      summary: Show learning configuration
      description: Returns values loaded from app.learning.* in application.yml.
      operationId: learningInfo
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Configuration loaded successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LearningInfoResponse"
components:
  schemas:
    ErrorResponse:
      type: object
      description: Error payload returned when the API rejects a request.
      properties:
        message:
          type: string
          description: Human-readable error message
          example: Course with id 99 was not found
        hint:
          type: string
          description: A small suggestion for the next action
          example: Check the resource id and try again
    CreateCourseRequest:
      type: object
      description: Request body used to create a persistent course.
      properties:
        title:
          type: string
          description: Course title
          example: Spring Boot Security
          maxLength: 100
          minLength: 3
        level:
          type: string
          description: Course level
          enum:
          - BEGINNER
          - INTERMEDIATE
          - ADVANCED
          example: INTERMEDIATE
        durationInHours:
          type: integer
          format: int32
          description: Estimated duration in hours
          example: 8
          minimum: 1
        published:
          type: boolean
          description: Whether the course is published
          example: true
      required:
      - level
      - title
    CourseResponse:
      type: object
      description: Persistent course returned from the database-backed API.
      properties:
        id:
          type: integer
          format: int64
          description: Generated database id
          example: 1
        title:
          type: string
          description: Course title
          example: Spring Boot Security
        level:
          type: string
          description: Course level
          enum:
          - BEGINNER
          - INTERMEDIATE
          - ADVANCED
          example: INTERMEDIATE
        durationInHours:
          type: integer
          format: int32
          description: Estimated duration in hours
          example: 8
        published:
          type: boolean
          description: Whether the course is published
          example: true
    CreateStudentRequest:
      type: object
      description: Request body used to create a new student.
      properties:
        name:
          type: string
          description: Student name
          example: Lina
        age:
          type: integer
          format: int32
          description: Student age
          example: 23
        active:
          type: boolean
          description: Whether the student starts as active
          example: true
        subjects:
          type: array
          description: Topics that the student is studying
          example:
          - annotations
          - controller
          items:
            type: string
        scores:
          type: object
          additionalProperties:
            type: integer
            format: int32
          description: Example scores by subject
          example:
            java: 91
            spring: 89
    StudentResponse:
      type: object
      description: Response returned by the student API.
      properties:
        id:
          type: integer
          format: int64
          description: Generated student id
          example: 3
        name:
          type: string
          description: Student name
          example: Lina
        age:
          type: integer
          format: int32
          description: Student age
          example: 23
        active:
          type: boolean
          description: Whether the student is active
          example: true
        level:
          type: string
          description: Derived level based on age
          example: adult
        subjects:
          type: array
          description: Topics being studied
          example:
          - annotations
          - controller
          items:
            type: string
        scores:
          type: object
          additionalProperties:
            type: integer
            format: int32
          description: Scores by subject
          example:
            java: 91
            spring: 89
        totalScore:
          type: integer
          format: int32
          description: Total of all score values
          example: 180
        subjectSummary:
          type: string
          description: Readable summary of the subjects list
          example: "annotations, controller"
    LearningInfoResponse:
      type: object
      description: Typed configuration values exposed for learning purposes.
      properties:
        title:
          type: string
          description: Configured learning title
          example: Java + Spring Boot Learning Repo
        focusTopics:
          type: array
          description: Configured focus topics
          example:
          - java basics
          - spring mvc
          - jpa
          items:
            type: string